Extract Sum and Average extensions to separate partial class files
