+++ modules/platforms/dotnet/Apache.Ignite/Sql/IgniteQueryableExtensions.cs
@@ -34,7 +34,7 @@ using Table;
 /// Use <see cref="IRecordView{T}.AsQueryable"/> and <see cref="IKeyValueView{TK,TV}.AsQueryable"/> to query Ignite with LINQ, then
 /// materialize the query results by calling <see cref="ToResultSetAsync{T}"/>.
 /// </summary>
-public static class IgniteQueryableExtensions
+public static partial class IgniteQueryableExtensions
 {
     /// <summary>
     /// Executes the Ignite query represented by the provided <paramref name="queryable"/> and returns the resulting
@@ -477,596 +477,6 @@ public static class IgniteQueryableExtensions
 
         var provider = queryable.ToQueryableInternal().Provider;
         return await provider.ExecuteSingleAsync<TResult>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<int> SumAsync(this IQueryable<int> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<int>, int>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<int>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<int> SumAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, int>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, int>>, int>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<int>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<int?> SumAsync(this IQueryable<int?> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<int?>, int?>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<int?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<int?> SumAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, int?>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, int?>>, int?>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<int?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<long> SumAsync(this IQueryable<long> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<long>, long>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<long>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<long> SumAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, long>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, long>>, long>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<long>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<long?> SumAsync(this IQueryable<long?> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<long?>, long?>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<long?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<long?> SumAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, long?>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, long?>>, long?>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<long?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<float> SumAsync(this IQueryable<float> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<float>, float>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<float>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<float> SumAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, float>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, float>>, float>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<float>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<float?> SumAsync(this IQueryable<float?> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<float?>, float?>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<float?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<float?> SumAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, float?>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, float?>>, float?>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<float?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<double> SumAsync(this IQueryable<double> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<double>, double>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<double> SumAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, double>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, double>>, double>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<double?> SumAsync(this IQueryable<double?> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<double?>, double?>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<double?> SumAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, double?>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, double?>>, double?>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<decimal> SumAsync(this IQueryable<decimal> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<decimal>, decimal>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<decimal>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<decimal> SumAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, decimal>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, decimal>>, decimal>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<decimal>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<decimal?> SumAsync(this IQueryable<decimal?> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<decimal?>, decimal?>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<decimal?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the sum of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the sum of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Sum", typeof(Queryable))]
-    public static async Task<decimal?> SumAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, decimal?>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, decimal?>>, decimal?>(Queryable.Sum).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<decimal?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the average of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the average of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Average`1", typeof(Queryable))]
-    public static async Task<double> AverageAsync(this IQueryable<int> queryable)
-    {
-        // TODO: With double, decimal (int, long, double -> double; float -> float; decimal -> decimal)
-        // TODO: With nullables
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<int>, double>(Queryable.Average).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the average of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the maximum value in the sequence.
-    /// </returns>
-    [DynamicDependency("Average`1", typeof(Queryable))]
-    public static async Task<double> AverageAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, int>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, int>>, double>(Queryable.Average).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the average of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the average of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Average`1", typeof(Queryable))]
-    public static async Task<double?> AverageAsync(this IQueryable<int?> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<int?>, double?>(Queryable.Average).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the average of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the maximum value in the sequence.
-    /// </returns>
-    [DynamicDependency("Average`1", typeof(Queryable))]
-    public static async Task<double?> AverageAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, int?>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, int?>>, double?>(Queryable.Average).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the average of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the average of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Average`1", typeof(Queryable))]
-    public static async Task<double> AverageAsync(this IQueryable<long> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<long>, double>(Queryable.Average).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the average of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the maximum value in the sequence.
-    /// </returns>
-    [DynamicDependency("Average`1", typeof(Queryable))]
-    public static async Task<double> AverageAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, long>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, long>>, double>(Queryable.Average).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the average of a sequence of values.
-    /// </summary>
-    /// <param name="queryable">Query.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the average of a sequence of values.
-    /// </returns>
-    [DynamicDependency("Average`1", typeof(Queryable))]
-    public static async Task<double?> AverageAsync(this IQueryable<long?> queryable)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<long?>, double?>(Queryable.Average).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression);
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double?>(expression).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Returns the average of a sequence of values.
-    /// </summary>
-    /// <typeparam name="TSource">Element type.</typeparam>
-    /// <param name="queryable">Query.</param>
-    /// <param name="selector">A projection function to apply to each element.</param>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.
-    /// The task result contains the maximum value in the sequence.
-    /// </returns>
-    [DynamicDependency("Average`1", typeof(Queryable))]
-    public static async Task<double?> AverageAsync<TSource>(
-        this IQueryable<TSource> queryable,
-        Expression<Func<TSource, long?>> selector)
-    {
-        IgniteArgumentCheck.NotNull(queryable, nameof(queryable));
-
-        var method = new Func<IQueryable<TSource>, Expression<Func<TSource, long?>>, double?>(Queryable.Average).GetMethodInfo();
-        var expression = Expression.Call(null, method, queryable.Expression, Expression.Quote(selector));
-
-        var provider = queryable.ToQueryableInternal().Provider;
-        return await provider.ExecuteSingleAsync<double?>(expression).ConfigureAwait(false);
     }
 
     /// <summary>
